requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
awscli==1.25.88
boto3==1.24.87
botocore==1.27.87
awscli==1.25.88
boto3==1.24.87
botocore==1.27.87
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
colorama
constantly==15.1.0
cryptography==38.0.4
daphne==3.0.2
decorator==4.4.2
Django==4.1.7
django-storages==1.13.1
docutils==0.16
gunicorn==20.1.0
django-storages==1.13.1
docutils==0.16
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
imageio==2.22.1
imageio-ffmpeg==0.4.7
incremental==21.3.0
jmespath==1.0.1
moviepy==1.0.3
msgpack==1.0.4
numpy==1.23.3
Pillow==9.3.0
proglog==0.1.10
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
PyYAML==5.4.1
reportlab==3.6.12
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==4.2.0
toml==0.10.2
tqdm==4.64.1
Twisted==22.10.0
#twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
uvicorn==0.20.0
whitenoise==6.2.0
zope.interface==5.4.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
priority==2.0.0
django-referrer-policy==1.0
whitenoise==6.2.0
google-auth==2.16.0
